requirements.txt
streamlit
pytesseract
opencv-python
pillow
numpy
pandas